+++ src/types/ServiceMetadata.ts
@@ -19,6 +19,12 @@ export interface ServiceMetadata<T, K extends keyof T> {
     global?: boolean;
 
     /**
+     * Indicates if instance of this class must be created on each its request.
+     * Global option is ignored when this option is used.
+     */
+    transient?: boolean;
+
+    /**
      * Allows to setup multiple instances the different classes under a single service id string or token.
      */
     multiple?: boolean;